# source/api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is Process Wiz App.",
    "version": "1.0.0",
    "title": "Process Wiz"
  },
  "host": "api.pwdev.internal.nlightnconsulting.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/collection": {
      "get": {
        "tags": [
          "Collection"
        ],
        "description": "Get all collections including files.",
        "parameters": [
          {
            "in": "query",
            "name": "collection_name",
            "description": "To filter by collection name.",
            "required": false,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/CollectionResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      },
      "post": {
        "tags": [
          "Collection"
        ],
        "description": "Create a new collection.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Collection object that needs to be added.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCollection"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
    "/collection/{collection_id}": {
      "put": {
        "tags": [
          "Collection"
        ],
        "description": "Update a new collection.",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "description": "Id of the collection to be updated.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Collection object that needs to be updated.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCollection"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Collection"
        ],
        "description": "Delete a collection.",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "description": "Id of the collection to be updated.",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
    "/doc-chat/chat": {
      "post": {
        "tags": [
          "DocChat"
        ],
        "description": "API query doc chat.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to send for doc chat.",
            "required": true,
            "schema": {
              "properties": {
                "user_query": {
                  "type": "string"
                },
                "thread_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
    "/doc-guide/chat": {
      "post": {
        "tags": [
          "DocGuide"
        ],
        "description": "API query doc guide chat.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to send for doc guide chat.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DocGuideRequestObject"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
    "/ask-doc/chat": {
      "post": {
        "tags": [
          "AskDoc"
        ],
        "description": "API query ask doc chat.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to send for doc chat.",
            "required": true,
            "schema": {
              "properties": {
                "collection_id": {
                  "type": "string"
                },
                "user_query": {
                  "type": "string"
                },
                "thread_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
    "/threads": {
      "get": {
        "tags": [
          "Threads"
        ],
        "description": "API to get all threads by module and type.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "module",
            "description": "Name of the module",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "type",
            "description": "Name of the type in module",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "thread_id": {
                        "type": "string"
                      },
                      "section_id": {
                        "type": "number"
                      },
                      "module": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "thread_created_datetime": {
                        "type": "string"
                      },
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "chat_id": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "user",
                                "assistant"
                              ]
                            },
                            "content": {
                              "type": "string"
                            },
                            "chat_created_datetime": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/threads/{thread_id}": {
      "get": {
        "tags": [
          "Threads"
        ],
        "description": "API to get one thread data by thread id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "description": "Id of the thread",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "thread_id": {
                      "type": "string"
                    },
                    "section_id": {
                      "type": "number"
                    },
                    "module": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "thread_created_datetime": {
                      "type": "string"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "chat_id": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "user",
                              "assistant",
                              "action"
                            ]
                          },
                          "content": {
                            "type": "string"
                          },
                          "chat_created_datetime": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
    "/doc-analysis/chat": {
      "post": {
        "tags": [
          "DocAnalysis"
        ],
        "description": "API query doc analysis.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to send for doc analysis.",
            "required": true,
            "schema": {
              "properties": {
                "user_query": {
                  "type": "string"
                },
                "file_id": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Type in the *'Value'* input box below: **'Bearer &lt;JWT&gt;'**, where JWT is the token"
    },
    "organization-id": {
      "type": "apiKey",
      "in": "header",
      "name": "organization-id",
      "description": "Id of the organization"
    }
  },
  "definitions": {
    "File": {
      "type": "object",
      "properties": {
        "file_id": {
          "type": "number"
        },
        "file_name": {
          "type": "string"
        },
        "file_url": {
          "type": "string"
        },
        "file_upload_datetime": {
          "type": "string"
        }
      }
    },
...
    "Collection": {
      "type": "object",
      "properties": {
        "collection_id": {
          "type": "number"
        },
        "collection_name": {
          "type": "string"
        },
        "collection_description": {
          "type": "string"
        },
        "files": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/File"
          }
        }
      }
    },
...
    "CreateCollection": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "CollectionResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collection"
          }
        }
      }
    },
...
    "NormalFailureResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "DocGuideRequestObject": {
      "type": "object",
      "properties": {
        "user_query": {
          "type": "string"
        },
        "file_id": {
          "type": "integer"
        },
        "thread_id": {
          "type": "string"
        }
      }
    },
...
  }
}